+++ test/keypin/core_test.clj
@@ -419,3 +419,28 @@
   (is (thrown? IllegalArgumentException
         (kp-bar-baz {:bar {:baz -1234}} :zap))        "invalid value exists")
   (is (= 2345  (kp-foo-bar {} 2345))                  "value absent, but default specified ar argument"))
+
+
+(def source (atom {}))
+
+
+(defkey
+  {:source source}
+  src-foo-bar [:foo integer? "Value at :foo" {:parser ku/any->int}]
+  src-bar-baz [:bar pos?     "Value at :bar" {:parser ku/any->int}])
+
+
+(defkey
+  nosrc-foo-bar [:foo integer? "Value at :foo" {:parser ku/any->int}]
+  nosrc-bar-baz [:bar pos?     "Value at :bar" {:parser ku/any->int}])
+
+
+(deftest test-source
+  (testing "Source is specified"
+    (reset! source {:foo -99
+                    :bar 100})
+    (is (= -99 (src-foo-bar)))
+    (is (= 100 (src-bar-baz))))
+  (testing "Source not specified"
+    (is (thrown? IllegalStateException (nosrc-foo-bar)))
+    (is (thrown? IllegalStateException (nosrc-bar-baz)))))